requirements.txt
python-telegram-bot==14.9.0
requests==2.28.2
zipfile